Forest Owner Verification Contract
Validates and manages forest landowners
