+++ Clarinet.toml
@@ -5,13 +5,13 @@ authors = []
 telemetry = true
 cache_dir = './.cache'
 requirements = []
-[contracts.certification-management]
-path = 'contracts/certification-management.clar'
+[contracts.forest-owner-verification]
+path = 'contracts/forest-owner-verification.clar'
 clarity_version = 2
 epoch = 2.5
 
-[contracts.forest-owner-verification]
-path = 'contracts/forest-owner-verification.clar'
+[contracts.harvest-planning]
+path = 'contracts/harvest-planning.clar'
 clarity_version = 2
 epoch = 2.5
 [repl.analysis]
