--- tests/certification-management.test.ts
@@ -1,104 +1,0 @@
-import { describe, it, expect, beforeEach } from "vitest"
-
-describe("Certification Management Contract", () => {
-  let contractAddress
-  let alice, bob, deployer
-  
-  beforeEach(() => {
-    contractAddress = "ST1PQHQKV0RJXZFY1DGX8MNSNYVE3VGZJSRTPGZGM.certification-management"
-    alice = "ST1SJ3DTE5DN7X54YDH5D64R3BCB6A2AG2ZQ8YPD5"
-    bob = "ST2CY5V39NHDPWSXMW9QDT3HC3GD6Q6XX4CFRK9AG"
-    deployer = "ST1PQHQKV0RJXZFY1DGX8MNSNYVE3VGZJSRTPGZGM"
-  })
-  
-  describe("Certification Issuance", () => {
-    it("should issue certifications with valid data", () => {
-      const result = {
-        type: "ok",
-        value: 1, // certification ID
-      }
-      
-      expect(result.type).toBe("ok")
-      expect(typeof result.value).toBe("number")
-    })
-    
-    it("should only allow contract owner to issue certifications", () => {
-      const result = {
-        type: "err",
-        value: 400, // ERR_UNAUTHORIZED
-      }
-      
-      expect(result.type).toBe("err")
-      expect(result.value).toBe(400)
-    })
-    
-    it("should validate compliance scores", () => {
-      const validScore = 85
-      const invalidScore = 150
-      
-      expect(validScore).toBeLessThanOrEqual(100)
-      expect(invalidScore).toBeGreaterThan(100)
-    })
-  })
-  
-  describe("Certification Validation", () => {
-    it("should correctly identify valid certifications", () => {
-      const currentBlock = 1000
-      const expiryBlock = 1500
-      const status = "active"
-      
-      const isValid = status === "active" && expiryBlock > currentBlock
-      expect(isValid).toBe(true)
-    })
-    
-    it("should correctly identify expired certifications", () => {
-      const currentBlock = 1000
-      const expiryBlock = 900
-      const status = "active"
-      
-      const isValid = status === "active" && expiryBlock > currentBlock
-      expect(isValid).toBe(false)
-    })
-    
-    it("should correctly identify revoked certifications", () => {
-      const currentBlock = 1000
-      const expiryBlock = 1500
-      const status = "revoked"
-      
-      const isValid = status === "active" && expiryBlock > currentBlock
-      expect(isValid).toBe(false)
-    })
-  })
-  
-  describe("Certification Management", () => {
-    it("should allow renewal of certifications", () => {
-      const result = {
-        type: "ok",
-        value: true,
-      }
-      
-      expect(result.type).toBe("ok")
-      expect(result.value).toBe(true)
-    })
-    
-    it("should allow revocation of certifications", () => {
-      const result = {
-        type: "ok",
-        value: true,
-      }
-      
-      expect(result.type).toBe("ok")
-      expect(result.value).toBe(true)
-    })
-    
-    it("should update compliance scores", () => {
-      const result = {
-        type: "ok",
-        value: true,
-      }
-      
-      expect(result.type).toBe("ok")
-      expect(result.value).toBe(true)
-    })
-  })
-})
